Reverts commits for extract parse color value to helper
